corrected problem of comversion from us to mertic and Back.

+++ trunk/menuplanner/grails-app/domain/com/mp/domain/Recipe.groovy
@@ -56,7 +56,7 @@ class Recipe extends Item {
 
     String getTotalTimeValue() {
         if (totalTime) {
-            BigDecimal time = totalTime?.value?.toBigDecimal()
+            Float time = totalTime?.value?.toFloat()
             return time.toString()
         } else {
             return null
